filters, counter, reportss

+++ src/DefaultBundle/Controller/AdvertsController.php
@@ -4,9 +4,7 @@ namespace DefaultBundle\Controller;
 
 use DefaultBundle\Entity\Adverts;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
-use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Validator\ConstraintViolationList;
 
 
@@ -22,28 +20,157 @@ class AdvertsController extends Controller
      */
     public function indexAction()
     {
+
+        $query1 = $this->getDoctrine()->getManager()->createQuery("select a from DefaultBundle:Adverts a where a.district = 'Суворовский'");
+        $a1 = $query1->getResult();
+        $amount1 = count($a1);
+
+        $query2 = $this->getDoctrine()->getManager()->createQuery("select a from DefaultBundle:Adverts a where a.district = 'Коминтерновский'");
+        $a2 = $query2->getResult();
+        $amount2 = count($a2);
+
+        $query3 = $this->getDoctrine()->getManager()->createQuery("select a from DefaultBundle:Adverts a where a.district = 'Малиновский'");
+        $a3 = $query3->getResult();
+        $amount3 = count($a3);
+
+        $query4 = $this->getDoctrine()->getManager()->createQuery("select a from DefaultBundle:Adverts a where a.district = 'Приморский'");
+        $a4 = $query4->getResult();
+        $amount4 = count($a4);
+
+        $query5 = $this->getDoctrine()->getManager()->createQuery("select a from DefaultBundle:Adverts a where a.district = 'Киевский'");
+        $a5 = $query5->getResult();
+        $amount5 = count($a5);
+
         $em = $this->getDoctrine()->getManager();
         $adverts = $em->getRepository('DefaultBundle:Adverts')->findAll();
-        $arr=[];
-        foreach ($adverts as $advert){
-            $ar = [
-                'title' => $advert->getTitle(),
-                'adress' => $advert->getAdress(),
-                'roomCount' => $advert->getRoomCount(),
-                'district' => $advert->getDistrict(),
-                'description' => $advert->getDiscription(),
-                'floor' => $advert->getFloor(),
-                'pricePerMonth' => $advert->getPricePerMonth(),
-//                'endDateOfAuction' => $advert->getEndDateOfAuction()->format('d.m.Y'),
-                'dateOfRenting' => $advert->getDateOfRenting()->format('d.m.Y')
 
-            ];
-            array_push($arr,json_encode($ar, true));
+        $counters = $this->getCounters();
+
+        return $this->render('adverts/index.html.twig', array(
+            'adverts' => $adverts,
+            'amountOfAdverts' => $counters[2],
+            'amountOfUsers' => $counters[1],
+            'amountOfComments' => $counters[0],
+            'amount1' => $amount1,
+            'amount2' => $amount2,
+            'amount3' => $amount3,
+            'amount4' => $amount4,
+            'amount5' => $amount5,
+
+        ));
+    }
+
+    public function getCounters()
+    {
+        $em = $this->getDoctrine()->getManager();
+        $adverts = $em->getRepository('DefaultBundle:Adverts')->findAll();
+        $users = $em->getRepository('DefaultBundle:Users')->findAll();
+        $comments = $em->getRepository('DefaultBundle:CommentsToAdvert')->findAll();
+
+        $countA = 0;
+        foreach ($adverts as $item) {
+            $countA = $countA + 1;
+        }
+        $countU = 0;
+        foreach ($users as $item) {
+            $countU = $countU + 1;
+        }
+        $countC = 0;
+        foreach ($comments as $item) {
+            $countC = $countC + 1;
         }
 
-        $response = new JsonResponse($arr);
-        $response->headers->set('Content-Type', 'application/json');
-        return $response;
+        $counters = array($countC, $countU, $countA);
+
+        return $counters;
+    }
+
+    public function filter1Action()
+    {
+        $query = $this->getDoctrine()->getManager()->createQuery("select a from DefaultBundle:Adverts a where a.district = 'Суворовский'");
+        $adverts = $query->getResult();
+
+
+
+        $counters = $this->getCounters();
+
+        return $this->render('adverts/index.html.twig', array(
+            'adverts' => $adverts,
+            'amountOfAdverts' => $counters[2],
+            'amountOfUsers' => $counters[1],
+            'amountOfComments' => $counters[0],
+
+        ));
+    }
+
+    public function filter2Action()
+    {
+        $query = $this->getDoctrine()->getManager()->createQuery("select a from DefaultBundle:Adverts a where a.district = 'Коминтерновский'");
+        $adverts = $query->getResult();
+
+
+
+        $counters = $this->getCounters();
+
+        return $this->render('adverts/index.html.twig', array(
+            'adverts' => $adverts,
+            'amountOfAdverts' => $counters[2],
+            'amountOfUsers' => $counters[1],
+            'amountOfComments' => $counters[0],
+        ));
+    }
+
+    public function filter3Action()
+    {
+        $query = $this->getDoctrine()->getManager()->createQuery("select a from DefaultBundle:Adverts a where a.district = 'Малиновский'");
+        $adverts = $query->getResult();
+
+
+
+        $counters = $this->getCounters();
+
+        return $this->render('adverts/index.html.twig', array(
+            'adverts' => $adverts,
+            'amountOfAdverts' => $counters[2],
+            'amountOfUsers' => $counters[1],
+            'amountOfComments' => $counters[0],
+
+        ));
+    }
+
+    public function filter4Action()
+    {
+        $query = $this->getDoctrine()->getManager()->createQuery("select a from DefaultBundle:Adverts a where a.district = 'Приморский'");
+        $adverts = $query->getResult();
+        $counters = $this->getCounters();
+
+
+
+        return $this->render('adverts/index.html.twig', array(
+            'adverts' => $adverts,
+            'amountOfAdverts' => $counters[2],
+            'amountOfUsers' => $counters[1],
+            'amountOfComments' => $counters[0],
+
+        ));
+    }
+
+    public function filter5Action()
+    {
+        $query = $this->getDoctrine()->getManager()->createQuery("select a from DefaultBundle:Adverts a where a.district = 'Киевский'");
+        $adverts = $query->getResult();
+
+
+
+        $counters = $this->getCounters();
+
+        return $this->render('adverts/index.html.twig', array(
+            'adverts' => $adverts,
+            'amountOfAdverts' => $counters[2],
+            'amountOfUsers' => $counters[1],
+            'amountOfComments' => $counters[0],
+
+        ));
     }
 
     /**
@@ -52,31 +179,38 @@ class AdvertsController extends Controller
      */
     public function newAction(Request $request)
     {
-        $data = json_decode($request->getContent(), true);
         $advert = new Adverts();
-        $user = $this->getUser();
-//        $advert->setUser($user);//------------------- ид юзера??
-        $advert->setAdress($data['adress']);
-        $advert->setFloor($data['floor']);
-        $advert->setDistrict($data['district']);
-        $advert->setDiscription($data['description']);
-        $advert->setTitle($data['title']);
-        $advert->setRoomCount($data['roomCount']);
-        $advert->setPricePerMonth($data['pricePerMonth']);
-//        if (isset($data['endDateOfAuction'])) {
-//            $data['endDateOfAuction'] = new \DateTime($data['endDateOfAuction']);
-//            $advert->setEndDateOfAuction($data['endDateOfAuction']);
-//        }
-        if (isset($data['dateOfRenting'])) {
-            $data['dateOfRenting'] = new \DateTime($data['dateOfRenting']);
-            $advert->setDateOfRenting($data['dateOfRenting']);
-        }
-        $em = $this->getDoctrine()->getManager();
-        $em->persist($advert);
-        $em->flush();
+        $form = $this->createForm('DefaultBundle\Form\AdvertsType', $advert);
+        $form->handleRequest($request);
 
-        $resp = new Response("ok");//-----------------------респонс со стасум выполнения----добавить валидацию и вывод ошибки------
-        return $resp;
+        if ($form->isSubmitted() && $form->isValid()) {
+            $errorList = $this->get('validator')->validate($advert);
+            if ($errorList->count() > 0)
+            {
+                foreach ($errorList as $error) {
+                    $this->addFlash('error', $error->getMessage());
+                }
+                return $this->redirectToRoute("adverts_new");
+            }
+
+            $userId = $this->getUser()->getId();
+
+            $em = $this->getDoctrine()->getManager();
+            $user = $em->getRepository('DefaultBundle:Users')->find($userId);
+
+            $advert->setUser($user);
+
+            $em = $this->getDoctrine()->getManager();
+            $em->persist($advert);
+            $em->flush();
+
+            return $this->redirectToRoute('adverts_show', array('id' => $advert->getId()));
+        }
+
+        return $this->render('adverts/new.html.twig', array(
+            'advert' => $advert,
+            'form' => $form->createView(),
+        ));
     }
 
     /**
@@ -85,69 +219,54 @@ class AdvertsController extends Controller
      */
     public function showAction(Adverts $advert)
     {
-       // $deleteForm = $this->createDeleteForm($advert);
+        $deleteForm = $this->createDeleteForm($advert);
 
-        $ar = [
-            'title' => $advert->getTitle(),
-            'adress' => $advert->getAdress(),
-            'roomCount' => $advert->getRoomCount(),
-            'district' => $advert->getDistrict(),
-            'description' => $advert->getDiscription(),
-            'floor' => $advert->getFloor(),
-            'pricePerMonth' => $advert->getPricePerMonth(),
-//            'endDateOfAuction' => $advert->getEndDateOfAuction()->format('d.m.Y'),
-            'dateOfRenting' => $advert->getDateOfRenting()->format('d.m.Y')
-
-        ];
-        $response = new Response(json_encode($ar));
-        $response->headers->set('Content-Type', 'application/json');
-        return $response;
-
+        return $this->render('adverts/show.html.twig', array(
+            'advert' => $advert,
+            'delete_form' => $deleteForm->createView(),
+        ));
     }
 
     /**
      * Displays a form to edit an existing advert entity.
      *
      */
-    public function editAction(Request $request, $id)
+    public function editAction(Request $request, Adverts $advert)
     {
-       $advert = $this->getDoctrine()->getManager()->getRepository('DefaultBundle:Adverts')->find($id);
-        $data = json_decode($request->getContent(), true);
-       // $user = $this->getUser();
-        $advert->setAdress($data['adress']);
-        $advert->setFloor($data['floor']);
-        $advert->setDistrict($data['district']);
-        $advert->setDiscription($data['description']);
-        $advert->setTitle($data['title']);
-        $advert->setRoomCount($data['roomCount']);
-        $advert->setPricePerMonth($data['pricePerMonth']);
-        if (isset($data['dateOfRenting'])) {
-            $data['dateOfRenting'] = new \DateTime($data['dateOfRenting']);
-            $advert->setDateOfRenting($data['dateOfRenting']);
-        }
-        $em = $this->getDoctrine()->getManager();
-        $em->persist($advert);
-        $em->flush();
+        $deleteForm = $this->createDeleteForm($advert);
+        $editForm = $this->createForm('DefaultBundle\Form\AdvertsType', $advert);
+        $editForm->handleRequest($request);
 
-        $resp = new Response("ok");//-----------------------респонс со стасум выполнения----добавить валидацию и вывод ошибки------
-        return $resp;
+        if ($editForm->isSubmitted() && $editForm->isValid()) {
+            $this->getDoctrine()->getManager()->flush();
+
+            return $this->redirectToRoute('adverts_edit', array('id' => $advert->getId()));
+        }
+
+        return $this->render('adverts/edit.html.twig', array(
+            'advert' => $advert,
+            'edit_form' => $editForm->createView(),
+            'delete_form' => $deleteForm->createView(),
+        ));
     }
 
     /**
      * Deletes a advert entity.
      *
      */
-    public function deleteAction($id)
+    public function deleteAction(Request $request, Adverts $advert)
     {
-        $advert = $this->getDoctrine()->getManager()->getRepository('DefaultBundle:Adverts')->find($id);
-        $em = $this->getDoctrine()->getManager();
-        $em->remove($advert);
-        $em->flush();
+        $form = $this->createDeleteForm($advert);
+        $form->handleRequest($request);
 
-        $resp = new Response("ok");//-----------------------респонс со стасум выполнения----добавить валидацию и вывод ошибки--
-        return $resp;
+        if ($form->isSubmitted() && $form->isValid()) {
+            $em = $this->getDoctrine()->getManager();
+            $em->remove($advert);
+            $em->flush();
+        }
 
-   }
+        return $this->redirectToRoute('adverts_index');
+    }
 
     /**
      * Creates a form to delete a advert entity.
@@ -156,12 +275,12 @@ class AdvertsController extends Controller
      *
      * @return \Symfony\Component\Form\Form The form
      */
-//    private function createDeleteForm(Adverts $advert)
-//    {
-//        return $this->createFormBuilder()
-//            ->setAction($this->generateUrl('adverts_delete', array('id' => $advert->getId())))
-//            ->setMethod('DELETE')
-//            ->getForm()
-//        ;
-//    }
+    private function createDeleteForm(Adverts $advert)
+    {
+        return $this->createFormBuilder()
+            ->setAction($this->generateUrl('adverts_delete', array('id' => $advert->getId())))
+            ->setMethod('DELETE')
+            ->getForm()
+        ;
+    }
 }
